src/functions_mqtt_ingest/requirements.txt
azure-functions==1.23.0
paho-mqtt==2.1.0
psycopg[binary]==3.2.9
